rid dead query executory and handler
