+++ core/src/main/scala/com/raphtory/internals/components/partition/QueryExecutor.scala
@@ -1,754 +1,273 @@
-//package com.raphtory.internals.components.partition
-//
-//import cats.effect.Async
-//import cats.effect.Resource
-//import cats.effect.std.Dispatcher
-//import cats.syntax.all._
-//import com.raphtory.api.analysis.graphstate.GraphState
-//import com.raphtory.api.analysis.graphstate.GraphStateImplementation
-//import com.raphtory.api.analysis.graphview._
-//import com.raphtory.api.analysis.table.Explode
-//import com.raphtory.api.analysis.table.Row
-//import com.raphtory.api.analysis.table.TableFilter
-//import com.raphtory.api.analysis.table.WriteToOutput
-//import com.raphtory.api.analysis.visitor.Vertex
-//import com.raphtory.api.output.sink.SinkExecutor
-//import com.raphtory.internals.communication.EndPoint
-//import com.raphtory.internals.communication.TopicRepository
-//import com.raphtory.internals.communication.SchemaProviderInstances._
-//import com.raphtory.internals.components.Component
-//import com.raphtory.internals.components.querymanager._
-//import com.raphtory.internals.graph.GraphPartition
-//import com.raphtory.internals.graph.LensInterface
-//import com.raphtory.internals.graph.Perspective
-//import com.raphtory.internals.management.Scheduler
-//import com.raphtory.internals.management.python.EmbeddedPython
-//import com.typesafe.config.Config
-//import com.typesafe.scalalogging.Logger
-//import org.slf4j.LoggerFactory
-//import java.util.concurrent.CompletableFuture
-//import java.util.concurrent.Semaphore
-//import java.util.concurrent.atomic.AtomicInteger
-//import java.util.concurrent.atomic.AtomicLong
-//import scala.collection.mutable
-//
-//private[raphtory] class QueryExecutor[F[_]](
-//    query: Query, // TODO we could use this directly instead of sending operations in every step
-//    partitionID: Int,
-//    storage: GraphPartition,
-//    conf: Config,
-//    topics: TopicRepository,
-//    scheduler: Scheduler,
-//    dispatcher: Dispatcher[F],
-//    finish: cats.effect.std.Semaphore[F]
-//) extends Component[QueryManagement](conf) {
-//
-//  private val graphID        = query.graphID
-//  private val sink           = query.sink.getOrElse(throw new Exception("No Sink found!"))
-//  private val jobID          = query.name
-//  private val pyScript       = query.pyScript
-//  private val logger: Logger = Logger(LoggerFactory.getLogger(this.getClass))
-//
-//  logger.info(logMessage(s"Starting QueryExecutor."))
-//
-//  private var currentPerspectiveID: Int        = _
-//  private var currentPerspective: Perspective  = _
-//  private var graphLens: LensInterface         = _
-//  private val sentMessageCount: AtomicLong     = new AtomicLong(0)
-//  private val receivedMessageCount: AtomicLong = new AtomicLong(0)
-//  private var votedToHalt: Boolean             = false
-//
-//  private val msgBatchPath: String  = "raphtory.partitions.batchMessages"
-//  private val messageBatch: Boolean = conf.getBoolean(msgBatchPath)
-//  private val maxBatchSize: Int     = conf.getInt("raphtory.partitions.maxMessageBatchSize")
-//
-//  private val sync = new QuerySuperstepSync(totalPartitions)
-//  pyScript.map(s => EmbeddedPython.global.run(s))
-//
-//  private val sinkExecutor: SinkExecutor = sink.executor(jobID, partitionID, conf)
-//
-//  if (messageBatch)
-//    logger.debug(
-//            logMessage(
-//                    s"Message batching is set to on. To change this modify '$msgBatchPath' in the application conf."
-//            )
-//    )
-//
-//  private val listener = topics.registerListener(
-//          s"$graphID-$jobID-query-executor-$partitionID",
-//          handleMessage,
-//          topics.jobOperations(graphID, jobID),
-//          partitionID
-//  )
-//  logger.debug(logMessage("Component message listener registered."))
-//
-//  private val vertexMessageListener =
-//    if (totalPartitions > 1)
-//      Some(
-//              topics.registerListener(
-//                      s"$graphID-$jobID-query-executor-$partitionID",
-//                      handleVertexMessages,
-//                      List(topics.vertexMessages(graphID, jobID), topics.vertexMessagesSync(graphID, jobID)),
-//                      partitionID
-//              )
-//      )
-//    else None
-//  logger.debug(logMessage("Vertex message listener registered."))
-//
-//  private val taskManager = topics.jobStatus(graphID, jobID).endPoint
-//  logger.debug(logMessage("TaskManager endpoint created"))
-//
-//  private val neighbours: Map[Int, EndPoint[VertexMessaging]] =
-//    if (totalPartitions > 1)
-//      topics.vertexMessages(graphID, jobID).endPoint()
-//    else
-//      Map.empty
-//  logger.debug(logMessage("Vertex message endpoints created"))
-//
-//  private val syncNeighbours: Map[Int, EndPoint[VertexMessagesSync]] = {
-//    if (totalPartitions > 1)
-//      topics.vertexMessagesSync(graphID, jobID).endPoint()
-//    else
-//      Map.empty
-//  }
-//  logger.debug(logMessage("Vertex control message endpoints created."))
-//
-//  override def run(): Unit = {
-//    logger.debug(logMessage(s"Starting query executor consumer."))
-//    listener.start()
-//    logger.debug(logMessage("Query executor consumer started, starting vertex message listeners."))
-//    vertexMessageListener.foreach(_.start())
-//    logger.debug(logMessage("Vertex message and vertex control message listeners started."))
-////    taskManager sendAsync ExecutorEstablished(partitionID)
-//    logger.debug(logMessage("QueryExecutor initialised."))
-//  }
-//
-//  override def stop(): Unit = {
-//    listener.close()
-//    vertexMessageListener.foreach(_.close())
-//    logger.debug(logMessage(s"closing query executor consumer."))
-//    taskManager.close()
-//    neighbours.values.foreach(_.close())
-//    syncNeighbours.values.foreach(_.close())
-//  }
-//
-//  def receiveVertexMessage(msg: VertexMessaging): Unit =
-//    try msg match {
-//
-////      case VertexMessageBatch(msgBatch) =>
-////        logger.trace(
-////                logMessage(
-////                        s"Executing 'VertexMessageBatch', '[${msgBatch
-////                          .mkString(",")}]'."
-////                )
-////        )
-////        msgBatch.foreach(message => graphLens.receiveMessage(message))
-////        receivedMessageCount.addAndGet(msgBatch.length)
-////        sync.updateVertexMessageCount(msgBatch.length)
-//
-//      case msg: GenericVertexMessage[_] =>
-//        logger.trace(
-//                logMessage(
-//                        s"Executing 'VertexMessage', '$msg'."
-//                )
-//        )
-//        graphLens.receiveMessage(msg)
-//        receivedMessageCount.incrementAndGet()
-//        sync.updateVertexMessageCount(1)
-//    }
-//    catch {
-//      case e: Throwable =>
-//        errorHandler(e)
-//    }
-//
-//  def receiveVertexControlMessage(msg: VertexMessagesSync): Unit = {
-//    logger.debug(
-//            logMessage(
-//                    s"Received control message from ${msg.partitionID}, should receive ${msg.count} messages"
-//            )
-//    )
-//    sync.updateControlMessageCount(msg.count)
-//  }
-//
-//  def handleVertexMessages(msg: QueryManagement): Unit =
-//    msg match {
-//      case x: VertexMessaging    => receiveVertexMessage(x)
-//      case x: VertexMessagesSync => receiveVertexControlMessage(x)
-//    }
-//
-//  override def handleMessage(msg: QueryManagement): Unit = {
-//    val time = System.currentTimeMillis()
-//    try {
-//      msg match {
-//        case CreatePerspective(id, perspective)                                       =>
-//          currentPerspectiveID = id
-//          currentPerspective = perspective
-//          receivedMessageCount.set(0)
-//          sentMessageCount.set(0)
-//          sync.reset()
-//          refreshBuffers()
-//          graphLens =
-//            storage.lens(jobID, perspective.actualStart, perspective.actualEnd, 0, sendMessage, errorHandler, scheduler)
-//          taskManager sendAsync PerspectiveEstablished(currentPerspectiveID, graphLens.localNodeCount)
-//          logger.debug(
-//                  logMessage(
-//                          s"Created perspective at time '${perspective.timestamp}' with window '${perspective.window}'. in ${System
-//                            .currentTimeMillis() - time}ms"
-//                  )
-//          )
-//
-//        case SetMetaData(vertices)                                                    =>
-//          graphLens.setFullGraphSize(vertices)
-//          taskManager sendAsync MetaDataSet(currentPerspectiveID)
-//          logger.debug(
-//                  logMessage(
-//                          s"Meta Data set in ${System.currentTimeMillis() - time}ms"
-//                  )
-//          )
-//
-//        case GraphFunctionWithGlobalState(function, graphState)                       =>
-//          function match {
-//            case StepWithGraph(f)                               =>
-//              evalStepWithGraph(time, graphState, f)
-//
-//            case iwg: IterateWithGraph[Vertex] @unchecked       =>
-//              startStep()
-//              val fun =
-//                if (iwg.executeMessagedOnly)
-//                  graphLens.runMessagedGraphFunction(iwg.f, graphState)(_)
-//                else
-//                  graphLens.runGraphFunction(iwg.f, graphState)(_)
-//              fun {
-//                finaliseStep {
-//                  val sentMessages     = sentMessageCount.get()
-//                  val receivedMessages = receivedMessageCount.get()
-//                  votedToHalt = graphLens.checkVotes()
-//                  taskManager sendAsync
-//                    GraphFunctionCompleteWithState(
-//                            currentPerspectiveID,
-//                            partitionID,
-//                            receivedMessages,
-//                            sentMessages,
-//                            votedToHalt,
-//                            graphState
-//                    )
-//
-//                  logger.debug(
-//                          logMessage(
-//                                  s"Iterate function on graph with accumulators completed  in ${System
-//                                    .currentTimeMillis() - time}ms and sent '$sentMessages' messages with `executeMessageOnly` flag set to ${iwg.executeMessagedOnly}."
-//                          )
-//                  )
-//                }
-//              }
-//
-//            case SelectWithGraph(f)                             =>
-//              startStep()
-//              graphLens.executeSelect(f, graphState) {
-//                finaliseStep {
-//                  taskManager sendAsync TableBuilt(currentPerspectiveID)
-//                  logger.debug(
-//                          logMessage(
-//                                  s"Select executed on graph with accumulators in ${System
-//                                    .currentTimeMillis() - time}ms."
-//                          )
-//                  )
-//                }
-//              }
-//
-//            case esf: ExplodeSelectWithGraph[Vertex] @unchecked =>
-//              startStep()
-//              graphLens.explodeSelect(esf.f, graphState) {
-//                finaliseStep {
-//                  taskManager sendAsync TableBuilt(currentPerspectiveID)
-//                  logger.debug(
-//                          logMessage(
-//                                  s"ExplodeSelect executed on graph with accumulators in ${System
-//                                    .currentTimeMillis() - time}ms."
-//                          )
-//                  )
-//                }
-//              }
-//            case GlobalSelect(f)                                =>
-//              evalGlobalSelect(time, graphState, f)
-//          }
-//
-//        case MultilayerView(interlayerEdgeBuilder)                                    =>
-//          startStep()
-//          graphLens.explodeView(interlayerEdgeBuilder) {
-//            finaliseStep {
-//              val sentMessages     = sentMessageCount.get()
-//              val receivedMessages = receivedMessageCount.get()
-//              taskManager sendAsync
-//                GraphFunctionComplete(
-//                        currentPerspectiveID,
-//                        partitionID,
-//                        receivedMessages,
-//                        sentMessages
-//                )
-//
-//              logger
-//                .debug(logMessage(s"MultilayerView function finished in ${System
-//                  .currentTimeMillis() - time}ms and sent '$sentMessages' messages."))
-//            }
-//          }
-//
-//        case ReduceView(defaultMergeStrategy, mergeStrategyMap, aggregate)            =>
-//          startStep()
-//          graphLens.reduceView(defaultMergeStrategy, mergeStrategyMap, aggregate) {
-//            finaliseStep {
-//              val sentMessages     = sentMessageCount.get()
-//              val receivedMessages = receivedMessageCount.get()
-//
-//              taskManager sendAsync
-//                GraphFunctionComplete(
-//                        currentPerspectiveID,
-//                        partitionID,
-//                        receivedMessages,
-//                        sentMessages
-//                )
-//
-//              logger
-//                .debug(logMessage(s"MultilayerView function finished in ${System
-//                  .currentTimeMillis() - time}ms and sent '$sentMessages' messages."))
-//            }
-//          }
-//
-//        case Step(f: (Vertex => Unit) @unchecked)                                     =>
-//          evaluateStep(time, f)
-//
-//        case UndirectedView()                                                         =>
-//          startStep()
-//          graphLens.viewUndirected() {
-//            finaliseStep {
-//              val sentMessages     = sentMessageCount.get()
-//              val receivedMessages = receivedMessageCount.get()
-//
-//              taskManager sendAsync
-//                GraphFunctionComplete(
-//                        currentPerspectiveID,
-//                        partitionID,
-//                        receivedMessages,
-//                        sentMessages
-//                )
-//
-//              logger
-//                .debug(logMessage(s"UndirectedView function finished in ${System
-//                  .currentTimeMillis() - time}ms and sent '$sentMessages' messages."))
-//            }
-//          }
-//
-//        case DirectedView()                                                           =>
-//          startStep()
-//          graphLens.viewDirected() {
-//            finaliseStep {
-//              val sentMessages     = sentMessageCount.get()
-//              val receivedMessages = receivedMessageCount.get()
-//
-//              taskManager sendAsync
-//                GraphFunctionComplete(
-//                        currentPerspectiveID,
-//                        partitionID,
-//                        receivedMessages,
-//                        sentMessages
-//                )
-//
-//              logger
-//                .debug(logMessage(s"DirectedView function finished in ${System
-//                  .currentTimeMillis() - time}ms and sent '$sentMessages' messages."))
-//            }
-//          }
-//
-//        case Iterate(f: (Vertex => Unit) @unchecked, iterations, executeMessagedOnly) =>
-//          evaluateIterate(time, f, executeMessagedOnly)
-//
-//        case ReversedView()                                                           =>
-//          startStep()
-//          graphLens.viewReversed() {
-//            finaliseStep {
-//              val sentMessages     = sentMessageCount.get()
-//              val receivedMessages = receivedMessageCount.get()
-//
-//              taskManager sendAsync
-//                GraphFunctionComplete(
-//                        currentPerspectiveID,
-//                        partitionID,
-//                        receivedMessages,
-//                        sentMessages
-//                )
-//
-//              logger
-//                .debug(logMessage(s"ReversedView function finished in ${System
-//                  .currentTimeMillis() - time}ms and sent '$sentMessages' messages."))
-//            }
-//          }
-//
-//        case ClearChain()                                                             =>
-//          graphLens.clearMessages()
-//          taskManager sendAsync GraphFunctionComplete(currentPerspectiveID, partitionID, 0, 0)
-//          logger.debug(
-//                  logMessage(
-//                          s"Messages cleared on graph in ${System
-//                            .currentTimeMillis() - time}ms."
-//                  )
-//          )
-//
-//        case s: Select[Vertex] @unchecked                                             =>
-//          startStep()
-//          graphLens.executeSelect(s.f) {
-//            finaliseStep {
-//              taskManager sendAsync TableBuilt(currentPerspectiveID)
-//              logger.debug(
-//                      logMessage(
-//                              s"Select executed on graph in ${System
-//                                .currentTimeMillis() - time}ms."
-//                      )
-//              )
-//            }
-//          }
-//
-//        //TODO create explode select with accumulators
-//        case es: ExplodeSelect[Vertex] @unchecked                                     =>
-//          startStep()
-//          graphLens.explodeSelect(es.f) {
-//            finaliseStep {
-//              taskManager sendAsync TableBuilt(currentPerspectiveID)
-//              logger.debug(
-//                      logMessage(
-//                              s"Exploded Select executed on graph in ${System
-//                                .currentTimeMillis() - time}ms."
-//                      )
-//              )
-//            }
-//          }
-//
-//        case TableFilter(f)                                                           =>
-//          graphLens.filteredTable(f) {
-//            taskManager sendAsync TableFunctionComplete(currentPerspectiveID)
-//            logger.debug(
-//                    logMessage(
-//                            s"Table Filter executed on table in ${System
-//                              .currentTimeMillis() - time}ms."
-//                    )
-//            )
-//          }
-//
-//        case Explode(f)                                                               =>
-//          graphLens.explodeTable(f) {
-//            taskManager sendAsync TableFunctionComplete(currentPerspectiveID)
-//            logger.debug(
-//                    logMessage(
-//                            s"Table Explode executed on table in ${System
-//                              .currentTimeMillis() - time}ms."
-//                    )
-//            )
-//          }
-//
-//        case WriteToOutput                                                            =>
-//          sinkExecutor.setupPerspective(currentPerspective)
-//          val writer = row => sinkExecutor.threadSafeWriteRow(row)
-//          graphLens.writeDataTable(writer) {
-//            sinkExecutor.closePerspective()
-//            taskManager sendAsync TableFunctionComplete(currentPerspectiveID)
-//            logger.debug(
-//                    logMessage(
-//                            s"Writing Results executed on table in ${System
-//                              .currentTimeMillis() - time}ms. Results written to '${sink.getClass.getSimpleName}'."
-//                    )
-//            )
-//          }
-//
-//        case CompleteWrite                                                            =>
-//          sinkExecutor.close()
-//          logger.debug(
-//                  logMessage(
-//                          "Received 'CompleteWrite' message. Output writer was successfully closed."
-//                  )
-//          )
-//          taskManager sendAsync WriteCompleted
-//
-//        case EndQuery(jobID)                                                          =>
-//          logger.debug(logMessage("Received 'EndQuery' message. "))
-//          dispatcher.unsafeRunSync(finish.release)
-//      }
-//    }
-//    catch {
-//      case e: Throwable =>
-//        errorHandler(e)
-//    }
-//    logger.debug(logMessage(s"Handled message $msg in ${System.currentTimeMillis() - time}ms"))
-//  }
-//
-//  private def evalGlobalSelect(time: Long, graphState: GraphStateImplementation, f: GraphState => Row): Unit = {
-//    startStep()
-//    if (partitionID == 0)
-//      graphLens.executeSelect(f, graphState) {
-//        taskManager sendAsync TableBuilt(currentPerspectiveID)
-//        logger.debug(
-//                logMessage(
-//                        s"Global Select executed on graph with accumulators in ${System
-//                          .currentTimeMillis() - time}ms."
-//                )
-//        )
-//      }
-//    else {
-//      taskManager sendAsync TableBuilt(currentPerspectiveID)
-//      logger.debug(
-//              logMessage(
-//                      s"Global Select executed on graph with accumulators in ${System
-//                        .currentTimeMillis() - time}ms."
-//              )
-//      )
-//    }
-//  }
-//
-//  private def evalStepWithGraph(time: Long, graphState: GraphStateImplementation, f: (_, GraphState) => Unit): Unit = {
-//    startStep()
-//    graphLens.runGraphFunction(f, graphState) {
-//      finaliseStep {
-//        val sentMessages     = sentMessageCount.get()
-//        val receivedMessages = receivedMessageCount.get()
-//        taskManager sendAsync
-//          GraphFunctionCompleteWithState(
-//                  currentPerspectiveID,
-//                  partitionID,
-//                  receivedMessages,
-//                  sentMessages,
-//                  graphState = graphState
-//          )
-//
-//        logger.debug(
-//                logMessage(
-//                        s"Step function on graph with accumulators finished in ${System
-//                          .currentTimeMillis() - time}ms and sent '$sentMessages' messages."
-//                )
-//        )
-//      }
-//    }
-//  }
-//
-//  private def evaluateIterate(time: Long, f: Vertex => Unit, executeMessagedOnly: Boolean): Unit = {
-//    startStep()
-//    val fun =
-//      if (executeMessagedOnly)
-//        graphLens.runMessagedGraphFunction(f)(_)
-//      else
-//        graphLens.runGraphFunction(f)(_)
-//    fun {
-//      finaliseStep {
-//        val sentMessages     = sentMessageCount.get()
-//        val receivedMessages = receivedMessageCount.get()
-//
-//        taskManager sendAsync
-//          GraphFunctionComplete(
-//                  currentPerspectiveID,
-//                  partitionID,
-//                  receivedMessages,
-//                  sentMessages,
-//                  graphLens.checkVotes()
-//          )
-//
-//        votedToHalt = graphLens.checkVotes()
-//        logger.debug(
-//                logMessage(
-//                        s"Iterate function completed in ${System
-//                          .currentTimeMillis() - time}ms and sent '$sentMessages' messages with `executeMessageOnly` flag set to $executeMessagedOnly."
-//                )
-//        )
-//      }
-//    }
-//  }
-//
-//  private def evaluateStep(time: Long, f: Vertex => Unit): Unit = {
-//    startStep()
-//    graphLens.runGraphFunction(f) {
-//      finaliseStep {
-//        val sentMessages     = sentMessageCount.get()
-//        val receivedMessages = receivedMessageCount.get()
-//        taskManager sendAsync
-//          GraphFunctionComplete(
-//                  currentPerspectiveID,
-//                  partitionID,
-//                  receivedMessages,
-//                  sentMessages
-//          )
-//
-//        logger
-//          .debug(logMessage(s"Step function finished in ${System
-//            .currentTimeMillis() - time}ms and sent '$sentMessages' messages."))
-//      }
-//    }
-//  }
-//
-//  def errorHandler(error: Throwable): Unit = {
-//    error.printStackTrace()
-//    taskManager sendAsync AlgorithmFailure(currentPerspectiveID, error)
-//  }
-//
-//  private val messageCache: Map[Int, mutable.ArrayBuffer[GenericVertexMessage[_]]] =
-//    neighbours.keys.map(n => n -> mutable.ArrayBuffer.empty[GenericVertexMessage[_]]).toMap
-//
-//  private val perStepSentMessageCounts: Map[Int, AtomicLong]                       = neighbours.keys.map(n => n -> new AtomicLong(0)).toMap
-//
-//  def sendMessage(message: GenericVertexMessage[_]): Unit = {
-//    val vId                  = message.vertexId match {
-//      case (v: Long, _) => v
-//      case v: Long      => v
-//    }
-//    sentMessageCount.incrementAndGet()
-//    val destinationPartition = (vId.abs % totalPartitions).toInt
-//    if (destinationPartition == partitionID) { //sending to this partition
-//      graphLens.receiveMessage(message)
-//      receivedMessageCount.incrementAndGet()
-//    }
-//    else { //sending to a remote partition
-//      perStepSentMessageCounts(destinationPartition).incrementAndGet()
-//      val producer =
-//        try neighbours(destinationPartition)
-//        catch {
-//          case e: NoSuchElementException =>
-//            val msg =
-//              s"Trying to send message to partition $destinationPartition from partition $partitionID but no endPoints were provided"
-//            logger.error(logMessage(msg))
-//            throw new IllegalStateException(msg)
-//        }
-//      if (messageBatch) {
-//        val cache = messageCache(destinationPartition)
-//        cache.synchronized {
-//          cache += message
-//          if (cache.size > maxBatchSize)
-//            sendCached(destinationPartition)
-//        }
-//      }
-//      else
-//        producer sendAsync message
-//    }
-//
-//  }
-//
-//  def sendCached(partition: Int): Unit = {
-//    val cache = messageCache(partition)
-//    cache.synchronized {
-////      neighbours(partition) sendAsync VertexMessageBatch(cache.toArray)
-//      cache.clear() // synchronisation breaks if we create a new object here
-//    }
-//  }
-//
-//  def flushMessages(): CompletableFuture[Void] = {
-//    logger.debug(logMessage("Flushing vertex messages."))
-//
-//    if (messageBatch)
-//      messageCache.keys.foreach(producer => sendCached(producer))
-//
-//    val futures = neighbours.values.map(_.flushAsync())
-//    CompletableFuture.allOf(futures.toSeq: _*)
-//  }
-//
-//  def flushControlMessages(): CompletableFuture[Void] = {
-//    logger.debug(logMessage("Flushing control messages."))
-//    CompletableFuture.allOf(syncNeighbours.values.map(_.flushAsync()).toSeq: _*)
-//  }
-//
-//  private def refreshBuffers(): Unit = {
-//    logger.debug(logMessage("Refreshing messageCache buffers for all Producers."))
-//    messageCache.values.foreach(_.clear())
-//  }
-//
-//  private def startStep(): Unit = {
-//    graphLens.nextStep()
-//    perStepSentMessageCounts.values.foreach(_.set(0))
-//  }
-//
-//  private def finaliseStep(f: => Unit): Unit = {
-//    syncNeighbours.foreach {
-//      case (id, endPoint) =>
-//        if (id != partitionID) {
-//          logger.debug(
-//                  logMessage(
-//                          s"Finished sending messages, final sent count for target $id is ${perStepSentMessageCounts(id).get()}"
-//                  )
-//          )
-//          endPoint.sendAsync(VertexMessagesSync(partitionID, perStepSentMessageCounts(id).get()))
-//        }
-//    }
-//    perStepSentMessageCounts.values.foreach(_.set(0))
-//    flushMessages()
-//      .thenCompose(_ => flushControlMessages())
-//      .thenCompose(_ => sync.awaitSuperstepComplete)
-//      .thenCompose(_ =>
-//        scheduler.executeCompletable {
-//          logger.debug(logMessage("Received all messages, finalising step"))
-//          f
-//        }
-//      )
-//  }
-//
-//  private def logMessage(msg: String): String = s"${jobID}_$partitionID: $msg"
-//}
-//
-//object QueryExecutor {
-//
-//  def apply[F[_]: Async](
-//      query: Query,
-//      partitionID: Int,
-//      storage: GraphPartition,
-//      conf: Config,
-//      topics: TopicRepository,
-//      scheduler: Scheduler
-//  ): F[Unit] =
-//    (for {
-//      dispatcher <- Dispatcher[F]
-//      finish     <- Resource.eval(cats.effect.std.Semaphore[F](0))
-//      _          <- makeExecutor(query, partitionID, storage, conf, topics, scheduler, dispatcher, finish)
-//    } yield finish).use(finish => finish.acquire) // I assume the finish.acquire implementation is defined as blocking
-//
-//  private def makeExecutor[F[_]: Async](
-//      query: Query,
-//      partitionID: Int,
-//      storage: GraphPartition,
-//      conf: Config,
-//      topics: TopicRepository,
-//      scheduler: Scheduler,
-//      dispatcher: Dispatcher[F],
-//      finish: cats.effect.std.Semaphore[F]
-//  ) =
-//    Resource
-//      .make(
-//              for {
-//                executor <-
-//                  Async[F].delay(
-//                          new QueryExecutor(query, partitionID, storage, conf, topics, scheduler, dispatcher, finish)
-//                  )
-//                _        <- Async[F].blocking(executor.run())
-//              } yield executor
-//      )(executor => Async[F].blocking(executor.stop()))
-//}
-//
-//class QuerySuperstepSync(totalPartitions: Int) {
-//  private val logger: Logger                             = Logger(LoggerFactory.getLogger(this.getClass))
-//  private val controlMessageSemaphore                    = new Semaphore(0)
-//  private val actualReceivedMessageCount: AtomicLong     = new AtomicLong(0)
-//  private val targetReceivedMessageCount: AtomicLong     = new AtomicLong(0)
-//  private val receivedControlMessageCount: AtomicInteger = new AtomicInteger(0)
-//
-//  def reset(): Unit = {
-//    actualReceivedMessageCount.set(0)
-//    targetReceivedMessageCount.set(0)
-//    receivedControlMessageCount.set(0)
-//  }
-//
-//  def updateControlMessageCount(count: Long): Unit = {
-//    targetReceivedMessageCount.addAndGet(count)
-//    controlMessageSemaphore.release(1)
-//  }
-//
-//  def updateVertexMessageCount(count: Int): Unit =
-//    actualReceivedMessageCount.addAndGet(count)
-//
-//  def awaitSuperstepComplete: CompletableFuture[Void] =
-//    CompletableFuture.runAsync { () =>
-//      controlMessageSemaphore.acquire(totalPartitions - 1)
-//      val target = targetReceivedMessageCount.get()
-//      while (!actualReceivedMessageCount.compareAndSet(target, target)) {}
-//    }
-//}
+package com.raphtory.internals.components.partition
+
+import cats.effect.Async
+import cats.effect.Ref
+import cats.effect.std.Dispatcher
+import cats.syntax.all._
+import com.google.protobuf.empty.Empty
+import com.raphtory.api.analysis.graphstate.GraphState
+import com.raphtory.api.analysis.graphstate.GraphStateImplementation
+import com.raphtory.api.analysis.graphview._
+import com.raphtory.api.analysis.table.Explode
+import com.raphtory.api.analysis.table.Row
+import com.raphtory.api.analysis.table.TableFilter
+import com.raphtory.api.analysis.table.TableFunction
+import com.raphtory.api.analysis.visitor.Vertex
+import com.raphtory.api.output.sink.SinkExecutor
+import com.raphtory.internals.components.querymanager.GenericVertexMessage
+import com.raphtory.internals.components.querymanager.Query
+import com.raphtory.internals.graph.GraphPartition
+import com.raphtory.internals.graph.LensInterface
+import com.raphtory.internals.graph.Perspective
+import com.raphtory.internals.management.Partitioner
+import com.raphtory.internals.management.Scheduler
+import com.raphtory.internals.serialisers.KryoSerialiser
+import com.raphtory.protocol._
+import com.typesafe.config.Config
+import com.typesafe.scalalogging.Logger
+import org.slf4j.LoggerFactory
+
+import scala.collection.mutable.ArrayBuffer
+
+class QueryExecutor[F[_]](
+    query: Query,
+    partitionID: Int,
+    storage: GraphPartition,
+    partitions: Map[Int, PartitionService[F]],
+    graphLens: Ref[F, Option[LensInterface]],
+    sinkExecutor: SinkExecutor,
+    messageBuffer: ArrayBuffer[GenericVertexMessage[_]],
+    dispatcher: Dispatcher[F],
+    conf: Config
+)(implicit F: Async[F]) {
+  private val logger: Logger = Logger(LoggerFactory.getLogger(this.getClass))
+  private val graphId        = query.graphID
+  private val jobId          = query.name
+  private val loggingPrefix  = s"${jobId}_$partitionID:"
+  private val kryo           = KryoSerialiser()
+  private val cb: () => Unit = () => {}
+
+  def receiveMessages(messages: Seq[GenericVertexMessage[_]]): F[Empty] =
+    withLens(lens => F.delay(messages.foreach(msg => lens.receiveMessage(msg)))).as(Empty())
+
+  def establishPerspective(perspective: Perspective): F[NodeCount] = {
+    val (timestamp, window) = (perspective.timestamp, perspective.window)
+    val (start, end)        = (perspective.actualStart, perspective.actualEnd)
+    timed(s"Creating perspective at time '$timestamp' with window '$window") {
+      for {
+        lens <- F.delay(storage.lens(jobId, start, end, 0, unsafeSendMessage, null, new Scheduler))
+        _    <- graphLens.update(_ => Some(lens))
+      } yield NodeCount(count = lens.localNodeCount)
+    }
+  }
+
+  def setMetadata(count: Int): F[Empty] =
+    timed(s"Setting metadata")(withLens(lens => F.delay(lens.setFullGraphSize(count)))).as(Empty())
+
+  def executeOperation(number: Int): F[OperationResult] = {
+    val function = query.operations(number)
+    timed(s"Executed operation $function") {
+      withLens { lens =>
+        function match {
+          case function: GraphFunction =>
+            F.delay(lens.nextStep()) *> executeGraphFunction(function, lens)
+          case function: TableFunction => // Table functions don't call nextStep
+            executeTableFunction(function, lens) as OperationResult(voteToContinue = true)
+        }
+      }
+    }
+  }
+
+  def executeOperationWithState(number: Int, state: GraphStateImplementation): F[OperationWithStateResult] = {
+    val function = query.operations(number).asInstanceOf[GraphFunction]
+    timed(s"Executed operation $function") {
+      withLens(lens => F.delay(lens.nextStep()) *> executeGraphFunctionWithState(function, state, lens))
+    }
+  }
+
+  def getResult: F[PartitionResult] =
+    timed(s"$loggingPrefix Getting results from table") {
+      withLens { lens =>
+        for {
+          buffer <- F.delay(ArrayBuffer[Row]())
+          // TODO Using kryo here is a hack so we get rows out of the row pool. Maybe we could consider if the pool mechanism is really necessary as it causes undesired behavior as in this case
+          _      <- F.blocking(lens.writeDataTable(row => buffer.addOne(kryo.deserialise[Row](kryo.serialise(row))))(cb))
+          result <- F.delay(PartitionResult(buffer.toSeq))
+        } yield result
+      }
+    }
+
+  def writePerspective(perspective: Perspective): F[Empty] =
+    timed(s"Writing results from table to sink ${query.sink.get.getClass.getSimpleName}") { // TODO unsafe call to get
+      withLens { lens =>
+        F.bracket(F.delay(sinkExecutor.setupPerspective(perspective))) { _ =>
+          F.blocking(lens.writeDataTable(row => sinkExecutor.threadSafeWriteRow(row))(cb))
+        }(_ => F.delay(sinkExecutor.closePerspective()))
+          .map(_ => Empty())
+      }
+    }
+
+  def endQuery: F[Empty] =
+    for {
+      _ <- F.delay(sinkExecutor.close())
+      _ <- F.delay(logger.debug(s"$loggingPrefix Received command to end the query"))
+    } yield Empty()
+
+  private def executeGraphFunction(function: GraphFunction, lens: LensInterface): F[OperationResult] = {
+    def processIterate(f: Vertex => Unit, executeMessagedOnly: Boolean) =
+      F.blocking(if (executeMessagedOnly) lens.runMessagedGraphFunction(f)(cb) else lens.runGraphFunction(f)(cb))
+
+    def processNonMessaging =
+      F.blocking(
+              function match {
+                case MultilayerView(interlayerEdgeBuilder) => lens.explodeView(interlayerEdgeBuilder)(cb)
+                case Step(f: (Vertex => Unit) @unchecked)  => lens.runGraphFunction(f)(cb)
+                case UndirectedView()                      => lens.viewUndirected()(cb)
+                case DirectedView()                        => lens.viewDirected()(cb)
+                case ReversedView()                        => lens.viewReversed()(cb)
+                case ClearChain()                          => lens.clearMessages(); cb()
+                case op: Select[Vertex] @unchecked         => lens.executeSelect(op.f)(cb)
+                case op: ExplodeSelect[Vertex] @unchecked  => lens.explodeSelect(op.f)(cb)
+                case rv: ReduceView                        => lens.reduceView(rv.defaultMergeStrategy, rv.mergeStrategyMap, rv.aggregate)(cb)
+                case x                                     => throw new Exception(s"$x not handled")
+              }
+      )
+
+    function match {
+      case Step(f: (Vertex => Unit) @unchecked) =>
+        wrapMessagingFunction(F.blocking(lens.runGraphFunction(f)(cb)), lens) as OperationResult()
+      case it: Iterate[Vertex] @unchecked       =>
+        for {
+          _      <- wrapMessagingFunction(processIterate(it.f, it.executeMessagedOnly), lens)
+          result <- F.blocking(OperationResult(voteToContinue = lens.checkVotes()))
+        } yield result
+      case _                                    => processNonMessaging as OperationResult()
+    }
+  }
+
+  private def executeGraphFunctionWithState(
+      function: GraphFunction,
+      graphState: GraphStateImplementation,
+      lens: LensInterface
+  ): F[OperationWithStateResult] = {
+    def processIterate(f: ((Vertex, GraphState) => Unit), executeMessagedOnly: Boolean) =
+      F.blocking(
+              if (executeMessagedOnly) lens.runMessagedGraphFunction(f, graphState)(cb)
+              else lens.runGraphFunction(f, graphState)(cb)
+      )
+
+    def processNonMessaging =
+      F.blocking(
+              function match {
+                case SelectWithGraph(f)                             => lens.executeSelect(f, graphState)(cb)
+                case esf: ExplodeSelectWithGraph[Vertex] @unchecked => lens.explodeSelect(esf.f, graphState)(cb)
+                case GlobalSelect(f)                                => if (partitionID == 0) lens.executeSelect(f, graphState)(cb) else cb()
+              }
+      )
+
+    function match {
+      case StepWithGraph(f)                        =>
+        for {
+          _ <- wrapMessagingFunction(F.blocking(lens.runGraphFunction(f, graphState)(cb)), lens)
+        } yield OperationWithStateResult(voteToContinue = true, graphState)
+      case it: IterateWithGraph[Vertex] @unchecked =>
+        for {
+          _      <- wrapMessagingFunction(processIterate(it.f, it.executeMessagedOnly), lens)
+          result <- F.blocking(OperationWithStateResult(lens.checkVotes(), graphState))
+        } yield result
+      case _                                       =>
+        processNonMessaging as OperationWithStateResult(voteToContinue = true, graphState)
+    }
+  }
+
+  private def executeTableFunction(function: TableFunction, lens: LensInterface): F[Unit] =
+    F.blocking(
+            function match {
+              case TableFilter(f) => lens.filteredTable(f)(cb)
+              case Explode(f)     => lens.explodeTable(f)(cb)
+            }
+    )
+
+  private def timed[A](processName: String)(f: F[A]) =
+    for {
+      output   <- F.timed(f)
+      (time, a) = output
+      _        <- F.delay(logger.debug(s"$loggingPrefix $processName took ${time.toMillis} ms to complete"))
+    } yield a
+
+  private def withLens[A](f: LensInterface => F[A]): F[A] = graphLens.get.flatMap(lens => f(lens.get))
+
+  private def wrapMessagingFunction(f: F[Unit], lens: LensInterface) = clearBuffer *> f *> sendMessages(lens)
+
+  private def sendMessages(lens: LensInterface) =
+    for {
+//      _           <- F.delay(println(s"in sendMessages for partition $partitionID"))
+      partitioned <- F.delay(messageBuffer.groupBy(message => partitionForMessage(message)))
+//      _           <- F.delay(println(s"messages partitioned for partition $partitionID"))
+      delivery     = partitioned.map {
+                       case (destination, messages) =>
+                         if (destination == partitionID)
+//                           F.delay(println(s"receiving direct messages $partitionID")) *>
+                           F.delay(messages.foreach(message => lens.receiveMessage(message)))
+                         else
+                           for {
+//                             _     <- F.delay(println(s"about to send chunk $partitionID"))
+                             chunk <- F.delay(messages.toSeq.asInstanceOf[Seq[GenericVertexMessage[Any]]])
+                             _     <- if (chunk.nonEmpty)
+                                        partitions(destination).receiveMessages(VertexMessages(graphId, jobId, chunk))
+                                      else F.unit
+                           } yield ()
+                     }
+//      _           <- F.delay(println(s"delivery defined for partition $partitionID"))
+      _           <- F.parSequenceN(partitions.size)(delivery.toSeq)
+    } yield ()
+
+  private def clearBuffer = F.delay(messageBuffer.clear())
+
+  private def unsafeSendMessage(message: GenericVertexMessage[_]): Unit =
+    // println(s"entering sync block for $partitionID")
+    messageBuffer.synchronized {
+      // println(s"inside sync block $partitionID")
+      messageBuffer.addOne(message)
+      if (messageBuffer.size >= 1024) {
+//        println(s"about to send chunk in $partitionID")
+        dispatcher.unsafeRunSync(for {
+//          _    <- F.delay(println(s"getting graphlens in $partitionID"))
+          lens <- graphLens.get
+//          _    <- F.delay(println(s"got graphlens in $partitionID"))
+          _    <- sendMessages(lens.get)
+        } yield ()) // graphLens.get.map(lens => sendMessages(lens.get)))
+//        println(s"messages sent in $partitionID")
+        messageBuffer.clear()
+//        println(s"buffer clear in $partitionID")
+      }
+      // println(s"getting out of syn block for $partitionID")
+    }
+
+  private def partitionForMessage(message: GenericVertexMessage[_]) = {
+    val vId = message.vertexId match {
+      case (v: Long, _) => v
+      case v: Long      => v
+    }
+    Partitioner().getPartitionForId(vId)
+  }
+}
+
+object QueryExecutor {
+
+  def apply[F[_]](
+      query: Query,
+      id: Int,
+      storage: GraphPartition,
+      partitions: Map[Int, PartitionService[F]],
+      dispatcher: Dispatcher[F],
+      conf: Config
+  )(implicit F: Async[F]): F[QueryExecutor[F]] =
+    for {
+      lens         <- Ref.of[F, Option[LensInterface]](None)
+      buffer       <- F.delay(ArrayBuffer[GenericVertexMessage[_]]())
+      sinkExecutor <- F.delay(query.sink.get.executor(query.name, id, conf)) // TODO: handle this as a resource
+      executor     <-
+        F.delay(new QueryExecutor[F](query, id, storage, partitions, lens, sinkExecutor, buffer, dispatcher, conf))
+    } yield executor
+}
